+++ CHANGELOG.md
@@ -3,6 +3,7 @@ In Development
 - Use `versioningit` for versioning
 - `maillog`: Fix a bug with deduplication of `tocc` contacts when inserting
   e-mails into the database
+- Added a global `--logfile` option for replacing stderr
 
 v2025.1.4
 ---------